refactor(tools): move the exec-denial half into denial.rs
Pure code motion, no behaviour change. `spawn_denied_path` and
`spawn_denied_hint` are the same concern as the write-denial code that just
landed in `tools/denial.rs` — translating a kernel EPERM that reached us
through a child's stderr — and were left in bash.rs only so that change stayed
reviewable as a behaviour change.

The moved functions, their doc comments, the `DENIED` fixture and both tests are
byte-identical to their previous form apart from `pub(super)` on the two
signatures; verified by hashing the extracted text before and after. bash.rs
drops to 453 lines and denial.rs sits at 421, both well inside the 800-line
limit that prompted the split.

The test count is unchanged at 940, which is the check that nothing was dropped
or duplicated in the move.
